Begin removal of name property from Player
There is no reason from a logical point of view why a player's name
should be part of the Player class, and so this patch begins to
migrate the property to Game_log instead. This should make it
easier to add extra non-logical properties (such as gender) to
players in the future.

This is a large-scale change, and this patch only partially enacts
it. The code WILL NOT COMPILE under this commit!